% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoCache.R
\name{addCacheType}
\alias{addCacheType}
\title{Adds a cache type to those available}
\usage{
addCacheType(type, f)
}
\arguments{
\item{type}{name of the cache type
f function that returns a list of functions bound to a cache name}
}
\value{
list of functions that can be used to manipulate the named cache
}
\description{
Adds a cache type to the registry of those available for initialisation.

A cache type is defined as a function that retuns a list of functions bound to a specific cache name as follows:
* set(key, value) - sets a value for a key, overwrites any existing values at that key
* get(key) - gets the value of a key, NULL otherwise
* unset(key) - unsets value of a key
* has(key) - true if the key has a value, false otherwise
* clear() - clears all values
}
\examples{

}
