% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.get}
\alias{cache.get}
\title{Cache Get}
\usage{
cache.get(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{value key}
}
\value{
value, NULL otherwise
}
\description{
Gets a value from the cache.
}
