% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.has}
\alias{cache.has}
\title{Cache Has}
\usage{
cache.has(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{value key}
}
\value{
TRUE if the cache contains the key, FALSE otherwise
}
\description{
Determines whether the key is in the cache.
}
