% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.set}
\alias{cache.set}
\title{Cache Set}
\usage{
cache.set(cache, key, value)
}
\arguments{
\item{cache}{cache}

\item{key}{value key}

\item{value}{value}
}
\value{
set value
}
\description{
Sets a value into the cache.
}
