% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoCache.R
\name{getCache}
\alias{getCache}
\title{Get initialised cache by name}
\usage{
getCache(name = "default")
}
\arguments{
\item{name}{name of the cache}
}
\value{
Named cache if initialised, otherwise error
}
\description{
Memoises a given function such that the result of the function is cached to improve
function performance
}
\examples{

}
