% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoCache.R
\name{hasCache}
\alias{hasCache}
\title{Has the named cache been initialised}
\usage{
hasCache(name = "default")
}
\arguments{
\item{name}{name of the cache}
}
\value{
True if the named cache has been initialised, false otherwise
}
\description{
Indicates whether the named cache has been initialised or not.
}
\examples{

}
