% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash.list}
\alias{hash.list}
\title{Hash}
\usage{
\method{hash}{list}(value)
}
\arguments{
\item{value}{value to hash}
}
\value{
hashed value as a string
}
\description{
Hashes a list of items, generating a single unique hash value which is based on the combination of hashed list items.
}
\examples{
my.function <- function (x, y) x+y

# a list of values to hash
values <- list(
  "Hello world!",
  101,
  3.142,
  TRUE,
  my.function,
  (function (x, y) x+y),
  functionCall(my.function, call("my.function", 10, 10)),
  list(a=1, b=2, c="hello")
)

# hash the values in the list
(hashes <- lapply(values, hash))

# Note that functions with the same body will have the same hash
hashes[[5]] == hashes[[6]]
}
