% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoCache.R
\name{initCache}
\alias{initCache}
\title{Get a reference to a named cache.}
\usage{
initCache(name = "default", type = "memory", algo = "sha1")
}
\arguments{
\item{name}{Cache name
type Type of cache
algo Hasing algorithm which defaults to sha1}
}
\value{

}
\description{
Gets a reference to the named cache, creating if one if it doesn't exist.
}
\examples{

}
