% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{is.memo}
\alias{is.memo}
\title{Is this a memo function}
\usage{
is.memo(f)
}
\arguments{
\item{f}{function, memo or otherwise}
}
\value{
\code{TRUE} if memo function, \code{FALSE} otherwise
}
\description{
Checks whether the passed function is a memo function.
}
