% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memo.cache}
\alias{memo.cache}
\title{Get memo function cache}
\usage{
memo.cache(f)
}
\arguments{
\item{f}{memo function}
}
\value{
Cache storing values for memoed function.
}
\description{
Gets the cache associated with a memo function allowing further manipulation and control of the underlying values
being stored.

Execution is stopped if function passed is not a valid memoed function.
}
