% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagefile.R
\name{storage.clear.file}
\alias{storage.clear.file}
\title{File Storage Clear}
\usage{
\method{storage.clear}{file}(cache)
}
\arguments{
\item{cache}{cache}
}
\description{
Clears the file storage of all values.
}
