% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagememory.R
\name{storage.clear.memory}
\alias{storage.clear.memory}
\title{Memory Storage Clear}
\usage{
storage.clear.memory(cache)
}
\arguments{
\item{cache}{cache}
}
\description{
Clears the memory storage of all values.
}
