% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.clear}
\alias{storage.clear}
\title{Clear the storage.}
\usage{
storage.clear(storage)
}
\arguments{
\item{storage}{initialized storage}
}
\value{
Invisibily returns storage
}
\description{
Clear the given storage of all keys and their values.
}
