% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagefile.R
\name{storage.get.file}
\alias{storage.get.file}
\title{File Storage Get}
\usage{
\method{storage.get}{file}(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}
}
\value{
value, NULL if not set
}
\description{
Gets a value from the file storage.
}
