% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagememory.R
\name{storage.get.memory}
\alias{storage.get.memory}
\title{Memory Storage Get}
\usage{
storage.get.memory(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}
}
\value{
value, NULL if not set
}
\description{
Gets a value from the memory storage.
}
