% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.get}
\alias{storage.get}
\title{Get value from a store.}
\usage{
storage.get(storage, key)
}
\arguments{
\item{storage}{initialized storage}

\item{key}{key to retrieve value for}
}
\value{
Stored value for the key, \code{NULL} otherwise.
}
\description{
Gets a value, for a given key, from the store.
 
If there is no coresponding value for the key, then \code{NULL} is returned.
}
