% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagefile.R
\name{storage.has.file}
\alias{storage.has.file}
\title{File Storage Has}
\usage{
\method{storage.has}{file}(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}
}
\value{
TRUE if the storage has the key, FALSE otherwise
}
\description{
Determines whether a key is set in the file storage.
}
