% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagememory.R
\name{storage.has.memory}
\alias{storage.has.memory}
\title{Memory Storage Has}
\usage{
storage.has.memory(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}
}
\value{
TRUE if the storage has the key, FALSE otherwise
}
\description{
Determines whether a key is set in the memory storage.
}
