% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagefile.R
\name{storage.set.file}
\alias{storage.set.file}
\title{File Storage Set}
\usage{
\method{storage.set}{file}(cache, key, value)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}

\item{value}{value}
}
\value{
set value
}
\description{
Sets a value into the file storage.
}
