% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagememory.R
\name{storage.set.memory}
\alias{storage.set.memory}
\title{Memory Storage Set}
\usage{
storage.set.memory(cache, key, value)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}

\item{value}{value}
}
\value{
set value
}
\description{
Sets a value into the memory storage.
}
