% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.set}
\alias{storage.set}
\title{Set value into a store.}
\usage{
storage.set(storage, key, value)
}
\arguments{
\item{storage}{initialized storage}

\item{key}{key to store value against}

\item{value}{value to store}
}
\value{
Invisbily returns storage
}
\description{
Stores a value for a given key.

If there is already a value stored for the key provided, then the exisiting value is 
overriden with the new value.
}
