% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagefile.R
\name{storage.unset.file}
\alias{storage.unset.file}
\title{File Storage Unset}
\usage{
\method{storage.unset}{file}(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}
}
\description{
Unsets a value from the file storage.
}
