% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storagememory.R
\name{storage.unset.memory}
\alias{storage.unset.memory}
\title{Memory Storage Unset}
\usage{
storage.unset.memory(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{key value}
}
\description{
Unsets a value from the memory storage.
}
