% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.unset}
\alias{storage.unset}
\title{Unset a value that corresponds to a key within a store.}
\usage{
storage.unset(storage, key)
}
\arguments{
\item{storage}{initialized storage}

\item{key}{key whose value is to be unset}
}
\value{
Invisibily returns storage
}
\description{
Unsets the value stored for a given key.

If there is no value for the key provided no action is taken.
}
