% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.get}
\alias{cache.get}
\title{Cache Get}
\usage{
cache.get(cache, key)
}
\arguments{
\item{cache}{cache}

\item{key}{value key}
}
\value{
value, NULL otherwise
}
\description{
Gets a value from the cache.
}
\examples{
# by default a memory cache will be created
myCache <- cache()

# the cache storage type can be specified, but only 'memory' is currently supported
alsoMyCache <- cache(storage = "memory")

# values can be stored in the cache 
cache.set(myCache, "name", "Pete Smith") 
cache.set(myCache, "age", "72")

# and retrieved
print(paste(cache.get(myCache, "name"), "is", cache.get(myCache, "age"), "years old."))

# setting a value with key that has already been used overwrites the currently stored value
cache.set(myCache, "name", "Sam Smith")
print(paste(cache.get(myCache, "name"), "is", cache.get(myCache, "age"), "years old."))

# set keys can be detected
if (cache.has(myCache, "name")) print("value for name key is set")

# keys can be unset
cache.unset(myCache, "name")
if (!cache.has(myCache, "name")) print("value for name key is not set")

# all values in the cache can be cleared
cache.clear(myCache)


     
}
