% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Create Cache}
\usage{
cache(storage = "memory")
}
\arguments{
\item{storage}{storage class}
}
\value{
a new cache
}
\description{
Creates a cache that can be used to store name, value pairs.

The provided storage class indicates where the values in the cache are stored for it's duration.
}
