% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Create a cache.}
\usage{
cache(id = NULL, storetype = "filestore")
}
\arguments{
\item{id}{Cache identifier, used to re-create reference to exisiting persistent cache.  If none provided then assumed transient.}

\item{storetype}{Persistant cache storage type.  File store by default.}
}
\value{
A cache that stores values for later retieval.
}
\description{
Create a general purpose cache that can be used to store both transient and persistent values.
}
