% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Simple cache used to store the values of memoised functions}
\usage{
cache()
}
\value{
A cache which can be accessed using:
\itemize{
  \item cache()#set - set the value of a given key
  \item cache()#get - get the value of a given key
  \item cache()#unset - clear the value for a given key
  \item cache()#has - true if given key has value, false otherwise
  \item cache()clear - clears all values and keys
  \item cache()#ls - lists all the values and keys 
  }
}
\description{
A simple in memory cache used to store the value of memoised functions.
}
