% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Get a cache}
\usage{
cache(algo = "sha1")
}
\arguments{
\item{algo}{hasing algorithm which defaults to sha1}
}
\value{
A cache that can be used to store values
}
\description{
Gets a cache of the type specified taking into account any context provided
}
