% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Create a cache.}
\usage{
cache(storage = NULL)
}
\arguments{
\item{storage}{Persistant cache storage. If none provided cache is transient and held in memory.  NULL by default.}
}
\value{
A cache that stores values for later retieval.
}
\description{
TODO
}
