% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoCache.R
\name{cache}
\alias{cache}
\title{Get a cache}
\usage{
cache(type = "memory", algo = "sha1")
}
\arguments{
\item{name}{Cache name
type Type of cache
algo Hasing algorithm which defaults to sha1}
}
\value{

}
\description{
Gets a cache of the type specified taking into account any context provided
}
\examples{

}
