% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash.functionCall}
\alias{hash.functionCall}
\title{Hash}
\usage{
\method{hash}{functionCall}(value)
}
\arguments{
\item{value}{value to hash}
}
\value{
hashed value as a string
}
\description{
Hashes a value into a string.
}
