% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memo.cache}
\alias{memo.cache}
\title{Get memo function cache}
\usage{
memo.cache(f)
}
\arguments{
\item{f}{memo function}
}
\value{
Cache storing values for memoed function.
}
\description{
Gets the cache associated with a memo function allowing further manipulation and control of the underlying values
being stored.

Execution is stopped if function passed is not a valid memoed function.
}
\examples{
# create a simple memo function
my.memo <- memo(function (someValue) {
  print("Getting some value")
  someValue
})

# make a couple of calls to memoed function to populate cache
my.memo(10)
my.memo(20)
my.memo(30)

# get the memo's cache
cache <- memo.cache(my.memo)

# clear the contents of the cache
cache$ls()     # expect to see cache entries from memo calls
cache$clear()
cache$ls()     # expect to see cache is now empty

# calling memo function again will repopulate cache
my.memo(10)
}
