% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memo.function}
\alias{memo.function}
\title{Memo Function}
\usage{
memo.function(f)
}
\arguments{
\item{f}{memo function}
}
\value{
Original unmemoized function.
}
\description{
Gets the original function that was memoized.

Execution is stopped if function passed is not a valid memoed function.
}
