% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memo}
\alias{memo}
\title{Memoises a given function}
\usage{
memo(f, allow.null = FALSE)
}
\arguments{
\item{f}{function to memoise}

\item{allow.null}{indicates whether results are cached when a function returns NULL, default value FALSE}
}
\value{
Memoised function
}
\description{
Memoises a given function such that the result of the function is cached to improve
function performance
}
