% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoFunction.R
\name{memo}
\alias{memo}
\title{Memoises a given function}
\usage{
memo(f, cache = NULL)
}
\arguments{
\item{f}{function to memoise}
}
\value{
Memoised function
}
\description{
Memoises a given function such that the result of the function is cached to improve
function performance
}
\examples{
# example function
example <- function (someValue) {
   Sys.sleep(someValue)
   somevalue
}

# memoise function
memoExample <- memo(example)

# call first time
memoExample(10)  # sleeps for 10 seconds and returns 10

# call second time
memoExample(10)  # immediately returns 10
}
