% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo.R
\name{memo}
\alias{memo}
\title{Memoises a given function}
\usage{
memo(f, cache)
}
\arguments{
\item{f}{function to memoise}

\item{cache}{cache to store functions memoised results in, if not specified creates a new in memory cache}
}
\value{
Memoised function
}
\description{
Memoises a given function such that the result of the function is cached to improve
function performance
}
\examples{
# example function
example <- function (someValue) {
  print("Getting some value")
  someValue
}

# memoise function
memoExample <- memo(example, cache())

# call first time
memoExample(10)  # prints "Getting some value" and returns 10

# call second time
memoExample(10)  # immediately returns
}
