% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memofunc.R
\docType{package}
\name{memofunc}
\alias{memofunc}
\alias{memofunc-package}
\title{memofunc: A package for memoizing functions and caching data}
\description{
The \code{memofunc} package provides a simple way to memoize a function to optimise execution for process or data
intensive actions.

\code{momefunc} also exposes a general purpose caching structure that can be used independanly.
}
\section{Functions}{

\itemize{
  \item \code{\link{memo}} - memoize a function
  \item \code{\link{is.memo}} - is the given function a memo
  \item \code{\link{memo.function}} - get a memo's origional function
  \item \code{\link{memo.cache}} - get a memo's cache
  \item \code{\link{cache}} - create a cache
}
}

\author{
Roy Wetherall \email{rwetherall@gmail.com}
}
