% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.clear}
\alias{storage.clear}
\title{Clear the storage.}
\usage{
storage.clear(storage)
}
\arguments{
\item{storage}{initialized storage}
}
\value{
Invisibily returns storage
}
\description{
Clear the given storage of all keys and their values.
}
\examples{
library(magrittr)

# initialize default memory storage
my.storage <- storage.init()

# set a value into storage
storage.set(my.storage, "name", "Roy Wetherall")

# .. and some more
my.storage \%>\% 
  storage.set("age", 45) \%>\% 
  storage.set("alive", TRUE) \%>\%
  storage.set("children", c("Peter", "Grace", "Lucy"))

# check a key has been set
if (storage.has(my.storage, "name")) print("I know your name!")

# .. and that a key hasn't been set
if (!storage.has(my.storage, "address")) print("I don't know where you live!")

# get some values from storage
sprintf(
  "\%s is \%i years old.", 
  storage.get(my.storage, "name"),
  storage.get(my.storage, "age"))
}
