% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.has}
\alias{storage.has}
\title{Has key has been used to store a value?}
\usage{
storage.has(storage, key)
}
\arguments{
\item{storage}{initialized storage}

\item{key}{key to check for stored value}
}
\value{
\code{TRUE} if key has an associated stored value, \code{FALSE} otherwise.
}
\description{
Indicates if a given key has a associated value stored in the storage or not.
}
