% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage.init}
\alias{storage.init}
\title{Initialize a store.}
\usage{
storage.init(storage.type = "memory", ...)
}
\arguments{
\item{storage.type}{storage type to initialize, defaults to \code{memory}}

\item{...}{additional configuration values used by storage implementations}
}
\value{
List containing characteristics perticular to the storage implementation, including:
\itemize{
   \item{\code{$type} - the storage type}
}
}
\description{
Initlaize storage for name value pairs based on provided type.

Available types of storage include:
\itemize{
   \item{\code{memory} - transient in-memory storage}
   \item{\code{file} - persistent storage, using local file storage}
}

Additional paramters may be provided when initializing different types of storage.  

See specific storage types for details.
}
